refactor: replace esutils.keywords.isIdentifierNameES6 by helper-validator-identifier
