+++ packages/babel-helper-builder-react-jsx-experimental/package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "@babel/helper-annotate-as-pure": "^7.8.3",
     "@babel/helper-module-imports": "^7.8.3",
-    "@babel/types": "^7.8.3",
-    "esutils": "^2.0.0"
+    "@babel/types": "^7.8.3"
   }
 }
