+++ packages/babel-helper-builder-react-jsx/package.json
@@ -10,7 +10,6 @@
   "main": "lib/index.js",
   "dependencies": {
     "@babel/helper-annotate-as-pure": "^7.8.3",
-    "@babel/types": "^7.8.3",
-    "esutils": "^2.0.0"
+    "@babel/types": "^7.8.3"
   }
 }
