+++ packages/babel-helper-validator-identifier/.npmignore
@@ -1,0 +1,3 @@
+src
+test
+*.log
