+++ packages/babel-helper-validator-identifier/scripts/generate-identifier-regex.js
