+++ packages/babel-types/package.json
@@ -14,7 +14,8 @@
   "dependencies": {
     "esutils": "^2.0.2",
     "lodash": "^4.17.13",
-    "to-fast-properties": "^2.0.0"
+    "to-fast-properties": "^2.0.0",
+    "@babel/helper-validator-identifier": "^7.9.0"
   },
   "devDependencies": {
     "@babel/generator": "^7.8.7",
